+++ ouroboros-network/sim-tests-lib/Test/Ouroboros/Network/Testnet/Internal.hs
@@ -73,6 +73,7 @@ import Network.TypedProtocol.PingPong.Type qualified as PingPong
 
 import Ouroboros.Network.ConnectionHandler (ConnectionHandlerTrace)
 import Ouroboros.Network.ConnectionManager.Core qualified as CM
+import Ouroboros.Network.ConnectionManager.State qualified as CM
 import Ouroboros.Network.ConnectionManager.Types (AbstractTransitionTrace)
 import Ouroboros.Network.ConsensusMode
 import Ouroboros.Network.Diffusion.P2P qualified as Diff.P2P
@@ -129,7 +130,6 @@ import Ouroboros.Network.PeerSelection.LocalRootPeers
            (OutboundConnectionsState (..))
 import Ouroboros.Network.PeerSelection.PeerAdvertise (PeerAdvertise (..))
 import Ouroboros.Network.PeerSelection.PeerSharing (PeerSharing)
-import Ouroboros.Network.PeerSelection.PeerTrustable (PeerTrustable)
 import Ouroboros.Network.PeerSelection.RelayAccessPoint (DomainAccessPoint (..),
            PortNumber, RelayAccessPoint (..))
 import Ouroboros.Network.PeerSelection.RootPeersDNS.DNSActions (DNSLookupType)
@@ -138,7 +138,7 @@ import Ouroboros.Network.PeerSelection.RootPeersDNS.LocalRootPeers
 import Ouroboros.Network.PeerSelection.RootPeersDNS.PublicRootPeers
            (TracePublicRootPeers)
 import Ouroboros.Network.PeerSelection.State.LocalRootPeers (HotValency (..),
-           WarmValency (..))
+           LocalRootConfig, WarmValency (..))
 import Ouroboros.Network.Protocol.PeerSharing.Codec (byteLimitsPeerSharing,
            timeLimitsPeerSharing)
 import Test.Ouroboros.Network.LedgerPeers (LedgerPools (..), genLedgerPoolsFrom)
@@ -206,8 +206,7 @@ data NodeArgs =
       -- ^ 'Arguments' 'aOwnPeerSharing' value
     , naLocalRootPeers         :: [( HotValency
                                    , WarmValency
-                                   , Map RelayAccessPoint ( PeerAdvertise
-                                                          , PeerTrustable)
+                                   , Map RelayAccessPoint LocalRootConfig
                                    )]
     , naLedgerPeers            :: Script LedgerPools
       -- ^ 'Arguments' 'LocalRootPeers' values
@@ -251,8 +250,7 @@ data Command = JoinNetwork DiffTime
              | Reconfigure DiffTime
                            [( HotValency
                             , WarmValency
-                            , Map RelayAccessPoint ( PeerAdvertise
-                                                   , PeerTrustable)
+                            , Map RelayAccessPoint LocalRootConfig
                             )]
   deriving Eq
 
@@ -268,7 +266,7 @@ instance Show Command where
 
 genCommands :: [( HotValency
                 , WarmValency
-                , Map RelayAccessPoint (PeerAdvertise, PeerTrustable)
+                , Map RelayAccessPoint LocalRootConfig
                 )]
             -> Gen [Command]
 genCommands localRoots = sized $ \size -> do
@@ -282,7 +280,7 @@ genCommands localRoots = sized $ \size -> do
   where
     subLocalRootPeers :: Gen [( HotValency
                               , WarmValency
-                              , Map RelayAccessPoint (PeerAdvertise, PeerTrustable)
+                              , Map RelayAccessPoint LocalRootConfig
                               )]
     subLocalRootPeers = do
       subLRP <- sublistOf localRoots
@@ -362,7 +360,7 @@ instance Arbitrary SmallPeerSelectionTargets where
 -- Simulation
 genNodeArgs :: [RelayAccessInfo]
             -> Int
-            -> [(HotValency, WarmValency, Map RelayAccessPoint (PeerAdvertise, PeerTrustable))]
+            -> [(HotValency, WarmValency, Map RelayAccessPoint LocalRootConfig)]
             -> RelayAccessInfo
             -> Gen NodeArgs
 genNodeArgs relays minConnected localRootPeers relay = flip suchThat hasUpstream $ do
@@ -681,7 +679,7 @@ genDiffusionScript :: ([RelayAccessInfo]
                         -> RelayAccessInfo
                         -> Gen [( HotValency
                                 , WarmValency
-                                , Map RelayAccessPoint (PeerAdvertise, PeerTrustable))])
+                                , Map RelayAccessPoint LocalRootConfig)])
                    -> RelayAccessInfosWithDNS
                    -> Gen (SimArgs, DomainMapScript, [(NodeArgs, [Command])])
 genDiffusionScript genLocalRootPeers
@@ -724,7 +722,7 @@ genNonHotDiffusionScript = genDiffusionScript genLocalRootPeers
                       -> RelayAccessInfo
                       -> Gen [( HotValency
                               , WarmValency
-                              , Map RelayAccessPoint (PeerAdvertise, PeerTrustable)
+                              , Map RelayAccessPoint LocalRootConfig
                               )]
     genLocalRootPeers relays _relay = flip suchThat hasUpstream $ do
       nrGroups <- chooseInt (1, 3)
@@ -757,7 +755,7 @@ genNonHotDiffusionScript = genDiffusionScript genLocalRootPeers
 
     hasUpstream :: [( HotValency
                     , WarmValency
-                    , Map RelayAccessPoint (PeerAdvertise, PeerTrustable)
+                    , Map RelayAccessPoint LocalRootConfig
                     )]
                 -> Bool
     hasUpstream localRootPeers =
@@ -784,7 +782,7 @@ genHotDiffusionScript = genDiffusionScript genLocalRootPeers
                         -> RelayAccessInfo
                         -> Gen [( HotValency
                                 , WarmValency
-                                , Map RelayAccessPoint (PeerAdvertise, PeerTrustable)
+                                , Map RelayAccessPoint LocalRootConfig
                                 )]
       genLocalRootPeers relays _relay = flip suchThat hasUpstream $ do
         let size = length relays
@@ -804,7 +802,7 @@ genHotDiffusionScript = genDiffusionScript genLocalRootPeers
 
       hasUpstream :: [( HotValency
                       , WarmValency
-                      , Map RelayAccessPoint (PeerAdvertise, PeerTrustable)
+                      , Map RelayAccessPoint LocalRootConfig
                       )]
                   -> Bool
       hasUpstream localRootPeers =
@@ -924,7 +922,7 @@ data DiffusionTestTrace =
           (ConnectionHandlerTrace NtNVersion NtNVersionData))
     | DiffusionDiffusionSimulationTrace DiffusionSimulationTrace
     | DiffusionConnectionManagerTransitionTrace
-        (AbstractTransitionTrace NtNAddr)
+        (AbstractTransitionTrace CM.ConnStateId)
     | DiffusionInboundGovernorTransitionTrace
         (IG.RemoteTransitionTrace NtNAddr)
     | DiffusionInboundGovernorTrace (IG.Trace NtNAddr)
@@ -994,7 +992,7 @@ diffusionSimulation
       :: Maybe ( Async m Void
                , StrictTVar m [( HotValency
                                , WarmValency
-                               , Map RelayAccessPoint (PeerAdvertise, PeerTrustable)
+                               , Map RelayAccessPoint LocalRootConfig
                                )])
          -- ^ If the node is running and corresponding local root configuration
          -- TVar.
@@ -1052,7 +1050,7 @@ diffusionSimulation
             -> Snocket m (FD m NtCAddr) NtCAddr
             -> StrictTVar m [( HotValency
                              , WarmValency
-                             , Map RelayAccessPoint (PeerAdvertise, PeerTrustable)
+                             , Map RelayAccessPoint LocalRootConfig
                              )]
             -> StrictTVar m (Map Domain [(IP, TTL)])
             -> m Void
@@ -1292,11 +1290,7 @@ diffusionSimulation
                                                        . tracerWithName ntnAddr
                                                        . tracerWithTime
                                                        $ nodeTracer
-        , Diff.P2P.dtConnectionManagerTransitionTracer = contramap
-                                                           DiffusionConnectionManagerTransitionTrace
-                                                       . tracerWithName ntnAddr
-                                                       . tracerWithTime
-                                                       $ nodeTracer
+        , Diff.P2P.dtConnectionManagerTransitionTracer = nullTracer
         , Diff.P2P.dtServerTracer                      = contramap DiffusionServerTrace
                                                        . tracerWithName ntnAddr
                                                        . tracerWithTime